requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
django-modeltranslation==0.19.12
django-phonenumber-field==8.0.0
Faker==33.3.1
phonenumbers==8.13.52
pillow==11.0.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
